[#35] refactor - 쿠폰 타입 개수 int 에서 long 으로 변경

+++ src/main/java/com/bamin/woorder/coupontype/domain/CouponType.java
@@ -59,7 +59,7 @@ public class CouponType extends ModifiableEntity {
         return couponTypeDiscount.getAmount();
     }
 
-    public int getCount() {
+    public Long getCount() {
         return couponTypeCount.getCount();
     }
